+++ build.txt
@@ -1,3 +1,3 @@
 displayName = Community Content Mod
 author = tModLoader Community
-version = 0.24
+version = 0.25
